add middleware validateBody
