+++ controllers/contacts.js
@@ -1,18 +1,5 @@
-const Joi = require("joi");
 const methods = require("../models");
 const { HttpError, ctrlWrapper } = require("../utils");
-
-const schema = Joi.object({
-  name: Joi.string().min(3).max(30).trim().required(),
-  email: Joi.string()
-    .email({
-      minDomainSegments: 2,
-      tlds: { allow: ["com", "net", "ua"] },
-    })
-    .trim()
-    .required(),
-  phone: Joi.string().min(9).max(15).trim().required(),
-});
 
 const listContacts = async (req, res, next) => {
   const result = await methods.listContacts();
@@ -30,22 +17,12 @@ const getContactById = async (req, res, next) => {
 };
 
 const addContact = async (req, res, next) => {
-  const { error } = schema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
-
   const result = await methods.addContact(req.body);
   return res.status(201).json(result);
 };
 
 const updateContact = async (req, res, next) => {
   const { contactId } = req.params;
-  const { error } = schema.validate(req.body);
-  if (error) {
-    throw HttpError(400, error.message);
-  }
-
   const result = await methods.updateContact(contactId, req.body);
 
   if (!result) {
