+++ routes/api/contacts.js
@@ -2,13 +2,16 @@ const express = require("express");
 const router = express.Router();
 const ctrl = require("../../controllers");
 
+const { validateBody } = require("../../middlewares");
+const { schema } = require("../../schemas/contacts");
+
 router.get("/", ctrl.listContacts);
 
 router.get("/:contactId", ctrl.getContactById);
 
-router.post("/", ctrl.addContact);
+router.post("/", validateBody(schema), ctrl.addContact);
 
-router.put("/:contactId", ctrl.updateContact);
+router.put("/:contactId", validateBody(schema), ctrl.updateContact);
 
 router.delete("/:contactId", ctrl.removeContact);
